Module Tjr_btree_examples

Examples for Tjr_btree

include Tjr_btree_examples__.Intf_

module type S = sig ... end

Input signature (types k and v, and related info)

type flg =

| O_TRUNC

| O_NOCACHE

Flags:

O_TRUNC, reinitialize B-tree root to point to an empty leaf; possibly also truncate the underlying file

O_NOCACHE, do not use a write back cache (the underlying file may still need to be sync'ed tho)

module type EX = sig ... end

Output signature (EX is "example")

Actual examples

module Make_example : sig ... end

Construct an example API, using lwt, bin_prot and bigarray

module Examples : sig ... end

Various examples

include Examples

module type INT_INT_EX = Tjr_btree_examples__.Intf_.EX with type k = int and type v = int and type t = Tjr_monad.lwt

module Int_int_ex = Examples.Int_int_ex