Module Tjr_btree_examples__Intf_

module type S = sig ... end

Input signature (types k and v, and related info)

type flg =

| O_TRUNC

| O_NOCACHE

Flags:

O_TRUNC, reinitialize B-tree root to point to an empty leaf; possibly also truncate the underlying file

O_NOCACHE, do not use a write back cache (the underlying file may still need to be sync'ed tho)

module type EX = sig ... end

Output signature (EX is "example")