Module Tjr_btree_examples__Store_read_cache

Add LRU caching to an existing store; two versions: a read cache, and a write back cache. FIXME superseded by fs_shared write_back_cache?

Read cache: writes are passed through in addition to being cached. So there is no need to flush anything. This dramatically improves performance.

Write cache: writes are buffered; multiple writes to the same block commit when LRU overflows (using the last write). Using a write back cache improves performance even more.

val add_imperative_read_cache_to_store : monad_ops:'a Tjr_monad.monad_ops -> store_ops:('blk_id, ('node, 'leaf) Isa_btree.dnode, 'a) Isa_btree.store_ops -> ('blk_id, ('node, 'leaf) Isa_btree.dnode, 'a) Isa_btree.store_ops

NOTE this adds an imperative read cache; take care with testing etc