Module Tjr_btree_examples

Examples for Tjr_btree

Introduction

There are two main types of example: in-memory and on-disk.

In-memory store

module Store_in_mem : sig ... end

Simple in-memory store implementation, mainly for testing

val mk_in_mem_store_ops : monad_ops:'a Tjr_monad.monad_ops -> mem_ops:('b, 'a) Store_in_mem.mem_ops -> (Store_in_mem.Blk_id.blk_id, 'b, 'a) Isa_btree.store_ops

Read cache and write-back cache based on LRU

module Store_read_cache : sig ... end

Add LRU caching to an existing store; two versions: a read cache, and a write back cache. FIXME superseded by fs_shared write_back_cache?

module Store_write_back_cache : sig ... end

NOTE alloc is necessary because a write of a node may only allocate the block, rather than actually writing.

Bin-prot marshalling for on-disk store

Use Jane St. marshalling libraries.

module Bin_prot_marshalling : sig ... end

Marshal frames to blocks using binprot. Currently assumes that blk_ids are ints.

Actual examples

module Make_example : sig ... end

Various examples, using lwt, bin_prot and bigarray

module Examples : sig ... end

Various examples

Internal

module Profilers = Tjr_btree_examples__.Profilers_