Module Tjr_btree_examples

Examples for Tjr_btree

module Intf_ : sig ... end

include Intf_

Interfaces based on classes and objects

class type ['k, 'v, 'ls] btree = object ... end

B-tree, with a write-back cache

class type ['k, 'v, 'ls] uncached_btree = object ... end

B-tree, no cache (write-back or read!)

class type open_fd = object ... end

Lwt-specific fd,blk_dev_ops

class type rt_blk = object ... end

Root block for simple examples NOTE: typically one of the init functions must be called

Refined construction

module Make_1 : sig ... end

Construct an example API, using lwt, bin_prot and bigarray and a write-back cache

module Make_2 : sig ... end

Wrapper around make_1, targeting an object

module Make_3 : sig ... end

Like Tjr_btree.Make, but using Std_types and binprot marshalling, and targetting an object; uncached

module Open_fd_and_rt_blk : sig ... end

Functionality to open a store, and handle the root block

Actual examples

module Int_int_ex = Make_1.Int_int_ex

module Generic_main_v2 : sig ... end

module Generic_example_v2 : sig ... end